+++ src/App.css
@@ -6,6 +6,7 @@
   margin: 0 auto;
   box-sizing: border-box;
   max-width: 100%;
+  padding: 0 20px;
 }
 
 main {
