+++ src/components/BurgerIngredients/burger-ingredients.module.css
@@ -7,12 +7,22 @@
 }
 .ingredientsChoise {
   display: flex;
-  flex-wrap: wrap;
+  flex-direction: column;
   justify-content: space-between;
   overflow-y: scroll;
   overflow-x: hidden;
   height: inherit;
+  width: 600px;
 }
 .ingredients h2 {
   margin-bottom: 20px;
+}
+.ingredients h3 {
+  width: 100%;
+  margin-top: 48px;
+  margin-bottom: 24px;
+}
+.ingidientsType {
+  display: flex;
+  flex-wrap: wrap;
 }
